move workaround to log.py
