+++ hubblestack/log.py
@@ -163,3 +163,17 @@ def emit_to_splunk(message, level, name):
     handler = SPLUNK_HANDLER
 
     handler.emit(MockRecord(message, level, time.asctime(), name))
+
+
+def workaround_salt_log_handler_queues():
+    class _FakeLogHandler(object):
+        level = 10
+        count = 0
+        def handle(self, record):
+            self.count += 1
+    flh = _FakeLogHandler()
+    import salt.log.setup as sls
+    sls.LOGGING_STORE_HANDLER.sync_with_handlers([flh])
+    sls.LOGGING_NULL_HANDLER.sync_with_handlers([flh])
+    # if flh.count > 0:
+    #     log.info("pretended to handle %d logging record(s) for salt.log.setup.LOGGING_*_HANDLER", flh.count)
